refactor : 코드 정리 및 병합
