--- src/main/java/com/grepp/team08/app/controller/web/image/ImageController.java
@@ -1,5 +1,0 @@
-package com.grepp.team08.app.controller.web.image;
-
-public class ImageController {
-
-}
