--- src/main/java/com/grepp/team08/app/controller/web/place/PlaceController.java
@@ -1,5 +1,0 @@
-package com.grepp.team08.app.controller.web.place;
-
-public class PlaceController {
-
-}
