--- src/main/java/com/grepp/team08/app/model/image/dto/ImageDto.java
@@ -1,5 +1,0 @@
-package com.grepp.team08.app.model.image.dto;
-
-public class ImageDto {
-
-}
