--- src/main/java/com/grepp/team08/app/model/place/service/PlaceService.java
@@ -1,5 +1,0 @@
-package com.grepp.team08.app.model.place.service;
-
-public class PlaceService {
-
-}
