--- src/main/java/com/grepp/team08/app/model/recommend/repository/RecommendRepository.java
@@ -1,5 +1,0 @@
-package com.grepp.team08.app.model.recommend.repository;
-
-public interface RecommendRepository {
-
-}
