--- src/main/java/com/grepp/team08/app/model/review/dto/ReviewDto.java
@@ -1,5 +1,0 @@
-package com.grepp.team08.app.model.review.dto;
-
-public class ReviewDto {
-
-}
