+++ src/main/resources/templates/admin_userManager.html
@@ -15,7 +15,7 @@
   <nav class="sidebar">
     <ul>
       <li><a th:href="@{/admin/users}">회원관리</a></li>
-      <li><a th:href="@{/admin/recommand-courses}">에디터픽코스</a></li>
+      <li><a th:href="@{/admin/recommend-courses}">에디터픽코스</a></li>
       <li><a th:href="@{/admin/editor-course}">에디터픽코스추가</a></li>
     </ul>
   </nav>
