+++ src/main/resources/templates/course_list.html
@@ -22,7 +22,7 @@
 <script>
   const courseList = document.getElementById('courseList');
 
-  fetch('/api/editor-recommand-courses')
+  fetch('/api/editor-recommend-courses')
   .then(response => {
     if (!response.ok) throw new Error("데이터를 불러오지 못했습니다");
     return response.json();
@@ -32,7 +32,7 @@
       const card = document.createElement('div');
       card.className = 'course-card';
       card.onclick = () => {
-        window.location.href = `/editor-recommand-courses/${course.courseId}`;
+        window.location.href = `/editor-recommend-courses/${course.courseId}`;
 
       };
