+++ src/main/resources/templates/mainPage.html
@@ -19,7 +19,7 @@
         <h2 class="section-title">에디터 픽 데이트 코스</h2>
         <div class="section-subtitle-row">
           <h2>에디터들이 엄선한 데이트 코스들이에요</h2>
-          <a th:href="@{/editor-recommand-courses}" class="more-link">더보기</a>
+          <a th:href="@{/editor-recommend-courses}" class="more-link">더보기</a>
         </div>
         <div class="carousel-wrapper">
           <div class="carousel" id="editor-carousel"></div>
