+++ src/test/java/com/grepp/team08/app/controller/api/place/PlaceApiControllerTest.java
@@ -4,7 +4,7 @@ import static org.junit.jupiter.api.Assertions.*;
 
 import org.junit.jupiter.api.Test;
 
-class MainPagePlaceApiControllerTest {
+class PlaceApiControllerTest {
 
   @Test
   void likeCourse() {
